requirements/local.txt
-r base.txt

mypy-extensions==0.4.3
mypy==0.991
black==19.10b0
isort==4.3.21
autoflake==1.7.8
flake8==3.9.2
pytest-cov==2.12.1
pytest==5.4.3
sqlalchemy-stubs==0.3
factory-boy==3.2.1
